Removed all of the useless BedWars classes
With a suggestion from @TheNullicorn this commit will delete all of the BedWars classes and refactor them into an Enum. This is because they we're simply holding references to the location that had to be found inside the JSON Object which could be held inside an enum.

The reason there are two classes, `HypixelBedWarsStats` and `HypixelBedWars` is because I wanted a way so each specific mode could not reference overall BedWars stats like the users level and coin amount.

The solution is not the most elegant, but it's the first idea I've got.
