--- src/main/java/io/github/hypixel_api_wrapper/wrapper/games/bedwars/AbstractHypixelBedWars.java
@@ -1,31 +1,0 @@
-package io.github.hypixel_api_wrapper.wrapper.games.bedwars;
-
-import io.github.hypixel_api_wrapper.wrapper.games.HypixelGame;
-import org.json.JSONObject;
-
-public abstract class AbstractHypixelBedWars extends HypixelGame {
-
-    protected AbstractHypixelBedWars(JSONObject stats) {
-        super(stats);
-    }
-
-    abstract String prefix();
-
-    @Override
-    public String getGameID() {
-        return "BEDWARS";
-    }
-
-    @Override
-    public String getGameName() {
-        return "Bed Wars";
-    }
-
-    @Override
-    public boolean isRemoved() {
-        return false;
-    }
-
-
-
-}
