--- src/main/java/io/github/hypixel_api_wrapper/wrapper/games/bedwars/HypixelBedWarsCastle.java
@@ -1,16 +1,0 @@
-package io.github.hypixel_api_wrapper.wrapper.games.bedwars;
-
-import org.json.JSONObject;
-
-public class HypixelBedWarsCastle extends AbstractHypixelBedWars {
-
-    public HypixelBedWarsCastle(JSONObject stats) {
-        super(stats);
-    }
-
-    @Override
-    String prefix() {
-        return "castle_";
-    }
-
-}
