--- src/main/java/io/github/hypixel_api_wrapper/wrapper/games/bedwars/HypixelBedWarsFours.java
@@ -1,17 +1,0 @@
-package io.github.hypixel_api_wrapper.wrapper.games.bedwars;
-
-import org.json.JSONObject;
-
-public class HypixelBedWarsFours extends AbstractHypixelBedWars {
-
-    public HypixelBedWarsFours(JSONObject stats) {
-        super(stats);
-    }
-
-    @Override
-    String prefix() {
-        return "four_four_";
-    }
-
-
-}
