--- src/main/java/io/github/hypixel_api_wrapper/wrapper/games/bedwars/HypixelBedWarsRushFours.java
@@ -1,16 +1,0 @@
-package io.github.hypixel_api_wrapper.wrapper.games.bedwars;
-
-import org.json.JSONObject;
-
-public class HypixelBedWarsRushFours extends AbstractHypixelBedWars {
-
-    public HypixelBedWarsRushFours(JSONObject stats) {
-        super(stats);
-    }
-
-    @Override
-    String prefix() {
-        return "four_four_rush_";
-    }
-
-}
