--- src/main/java/io/github/hypixel_api_wrapper/wrapper/games/bedwars/HypixelBedWarsSolos.java
@@ -1,17 +1,0 @@
-package io.github.hypixel_api_wrapper.wrapper.games.bedwars;
-
-import org.json.JSONObject;
-
-public class HypixelBedWarsSolos extends AbstractHypixelBedWars {
-
-    public HypixelBedWarsSolos(JSONObject stats) {
-        super(stats);
-    }
-
-    @Override
-    String prefix() {
-        return "eight_one_";
-    }
-
-
-}
