--- src/main/java/io/github/hypixel_api_wrapper/wrapper/games/bedwars/HypixelBedWarsTrios.java
@@ -1,16 +1,0 @@
-package io.github.hypixel_api_wrapper.wrapper.games.bedwars;
-
-import org.json.JSONObject;
-
-public class HypixelBedWarsTrios extends AbstractHypixelBedWars {
-
-    public HypixelBedWarsTrios(JSONObject stats) {
-        super(stats);
-    }
-
-    @Override
-    String prefix() {
-        return "four_three_";
-    }
-
-}
