--- src/main/java/io/github/hypixel_api_wrapper/wrapper/games/bedwars/HypixelBedWarsUltimateFours.java
@@ -1,17 +1,0 @@
-package io.github.hypixel_api_wrapper.wrapper.games.bedwars;
-
-import org.json.JSONObject;
-
-public class HypixelBedWarsUltimateFours extends AbstractHypixelBedWars {
-
-    public HypixelBedWarsUltimateFours(JSONObject stats) {
-        super(stats);
-    }
-
-    @Override
-    String prefix() {
-        return "four_four_ultimate_";
-    }
-
-
-}
