--- src/main/java/io/github/hypixel_api_wrapper/wrapper/hypixel_object_factory/HypixelObjectFactory.java
@@ -1,12 +1,0 @@
-package io.github.hypixel_api_wrapper.wrapper.hypixel_object_factory;
-
-import org.json.JSONObject;
-
-public abstract class HypixelObjectFactory {
-
-    JSONObject stats;
-
-    public HypixelObjectFactory(JSONObject stats) {
-        this.stats = stats;
-    }
-}
